pyproject.toml
[tool.poetry]
name = "pycenca-cal"
version = "0.1.0"
description = ""
authors = ["Jorge Calatayud <jorge.cmaeso@alumnos.upm.es>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "^1.21.1"
matplotlib = "^3.5.1"
PyVISA = "^1.11.3"
scipy = "^1.7.3"
loguru = "^0.6.0"
PyYAML = "^6.0"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
